+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   STAC Items ([#430](https://github.com/stac-utils/pystac/pull/430))
 - Support for Python 3.9 ([#420](https://github.com/stac-utils/pystac/pull/420))
 - Migration for pre-1.0.0-rc.1 Stats Objects (renamed to Range Objects in 1.0.0-rc.3) ([#447](https://github.com/stac-utils/pystac/pull/447))
+- Attempt to extend a `STACObject` that does not contain the extension's schema URI in
+  `stac_extensions` raises new `ExtensionNotImplementedError`.
 
 ### Changed